[great_wall_motor] Fix and rename spider
